Expose new functions in ESM wrapper
Some time ago coloris added the new functions `setInstance`, `removeInstance`, and `updatePosition`. We also added them in the type definitions. But attempting to use those functions when importing from the ESM modules gives the following error:

```
✘ [ERROR] No matching export in "../../.yarn/cache/@melloware-coloris-patch-884dd6decc-cd6a043de1.zip/node_modules/@melloware/coloris/dist/esm/coloris.js" for import "setInstance"
✘ [ERROR] No matching export in "../../.yarn/cache/@melloware-coloris-patch-884dd6decc-cd6a043de1.zip/node_modules/@melloware/coloris/dist/esm/coloris.js" for import "removeInstance"
✘ [ERROR] No matching export in "../../.yarn/cache/@melloware-coloris-patch-884dd6decc-cd6a043de1.zip/node_modules/@melloware/coloris/dist/esm/coloris.js" for import "updatePosition"
```

ESM requires all exports to be named explicitly, so they need to be added to the esm-wrapper.js